+++ tests/testthat/test-CBA.R
@@ -5,7 +5,7 @@ data("iris")
 context("CBA")
 
 cba_classifier <- CBA(Species ~ ., iris, supp = 0.05, conf = 0.9, pruning = "M1")
-expect_equal(length(rules(cba_classifier)), 7L)
+expect_equal(length(rules(cba_classifier)), 8L)
 expect_equal(cba_classifier$default, "Species=versicolor")
 
 results <- predict(cba_classifier, iris)
@@ -26,4 +26,7 @@ results <- predict(cba_classifier, head(iris, n = 5))
 expect_equal(length(results), 5L)
 
 # FIXME: We need to check what the output of M2 should be
-cba_classifier <- CBA(Species ~ ., iris, supp = 0.05, conf = 0.9, pruning = "M2")
+cba_classifier_M2 <- CBA(Species ~ ., iris, supp = 0.05, conf = 0.9, pruning = "M2")
+# FIXME: there is a bug in totalError calculation in M2
+#expect_equal(length(rules(cba_classifier_M2)), 8L)
+